rpc: command_runner: move header to public include
Move the `command_runner.h` file to the public include path so
downstream RPC implementations can use the functions.
