+++ subsys/rpc/commands/data_logger_read.c
@@ -12,9 +12,8 @@
 #include <infuse/data_logger/logger.h>
 #include <infuse/epacket/packet.h>
 #include <infuse/rpc/commands.h>
+#include <infuse/rpc/command_runner.h>
 #include <infuse/rpc/types.h>
-
-#include "../command_runner.h"
 
 struct net_buf *rpc_command_data_logger_read(struct net_buf *request)
 {
