+++ subsys/rpc/commands/data_receiver.c
@@ -11,10 +11,9 @@
 #include <zephyr/sys/crc.h>
 
 #include <infuse/rpc/commands.h>
+#include <infuse/rpc/command_runner.h>
 #include <infuse/rpc/types.h>
 #include <infuse/epacket/packet.h>
-
-#include "../command_runner.h"
 
 LOG_MODULE_DECLARE(rpc_server);
