+++ subsys/rpc/commands/file_write_basic.c
@@ -14,6 +14,7 @@
 
 #include <infuse/dfu/helpers.h>
 #include <infuse/rpc/commands.h>
+#include <infuse/rpc/command_runner.h>
 #include <infuse/rpc/types.h>
 #include <infuse/epacket/packet.h>
 #include <infuse/reboot.h>
@@ -24,7 +25,6 @@
 #endif /* CONFIG_NRF_MODEM_LIB */
 
 #include "common_file_actions.h"
-#include "../command_runner.h"
 
 LOG_MODULE_DECLARE(rpc_server);
